chore: simplify returns a little
