+++ src/components/pages/Executors/ExecutorDetails/ExecutorSettings/CommandAndArguments/CommandAndArguments.tsx
@@ -1,13 +1,11 @@
 import Arguments from './Arguments';
 import Command from './Command';
 
-const CommandAndArguments: React.FC = () => {
-  return (
-    <>
-      <Command />
-      <Arguments />
-    </>
-  );
-};
+const CommandAndArguments: React.FC = () => (
+  <>
+    <Command />
+    <Arguments />
+  </>
+);
 
 export default CommandAndArguments;
