+++ src/components/pages/Tests/TestDetails/TestSettings/SettingsExecution/SettingsExecution.tsx
@@ -1,13 +1,11 @@
 import PostRun from './PostRun';
 import PreRun from './PreRun';
 
-const SettingsExecution: React.FC = () => {
-  return (
-    <>
-      <PreRun />
-      <PostRun />
-    </>
-  );
-};
+const SettingsExecution: React.FC = () => (
+  <>
+    <PreRun />
+    <PostRun />
+  </>
+);
 
 export default SettingsExecution;
